+++ .eslintrc.js
@@ -1,6 +1,3 @@
 module.exports = {
-  extends: [require.resolve('@umijs/fabric/dist/eslint')],
-  rules: {
-    // your rules
-  },
+  extends: require.resolve('@umijs/lint/dist/config/eslint'),
 };