+++ src/index.ts
@@ -1,0 +1,2 @@
+export { default as SakanaReact } from './SakanaReact';
+export * from './hooks';